Game mechanics

+++ core/src/com/ajeffcorrigan/game/animalmatch/gamesystem/GameCell.java
@@ -99,6 +99,6 @@ public class GameCell {
     }
 
     public int getTileCode(int size) {
-        return (int)logicCoordinates.x + ((int)logicCoordinates.y * (int)size);
+        return (int)logicCoordinates.x + ((int)logicCoordinates.y * size);
     }
 }
